extract test helpers
